+++ base_cmd.vitis/WAVE_TestSSD/src/nvme.c
@@ -335,9 +335,9 @@ int nvmeServiceIOCompletions(u16 maxCompletions)
 	return numCompletions;
 }
 
-u16 nvmeGetIOSlip(void)
+int nvmeGetIOSlip(void)
 {
-	return (u16)(io_cid - io_cid_last_completed - 1);
+	return ((int)io_cid - (int)io_cid_last_completed - 1);
 }
 
 // Private Function Definitions ----------------------------------------------------------------------------------------
